Changed 'selection' of rows to DATETIME, and 'position' to FLOAT.
This way, it is possible to update a single row to make it selected (i.e., set it to the current time) or to reposition it among peers (set it to the average of the before and after row's positions).

+++ database/src/main/resources/my-sql/300-project_milestone-table.sql
@@ -4,8 +4,8 @@ CREATE TABLE project_milestone
     creation_time   DATETIME     NOT NULL DEFAULT CURRENT_TIMESTAMP,
     deleted         DATETIME,
 
-    -- Indicates the relative location on the milestone.
-    position        INT          NOT NULL,
+    -- Indicates the relative position among a project's milestones.
+    position        FLOAT        NOT NULL,
 
     name            VARCHAR(255) NOT NULL,
     short_descr     TEXT,
